kernel: Boost DDR bus for a short amount of time when zygote forks
There is a gap from when a new app is forked to the time that the
devfreq governor managing the DDR bus can pick up the increased load and
ramp up appropriately. Since the polling interval for this governor is
typically 50 ms, that means it may take up to 50 ms before the governor
ramps up, which hurts cold launch app performance. Boosting the DDR bus
for 50 ms upon a fork from zygote covers that gap, speeding up cold app
launches.

Change-Id: I57a26a9f6cead3c60c7e94a6d6c49e50ef2e7f75
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>