LRNG: allow nonblocking interfaces when LRNG is non-operational
When the LRNG is turned non-operational, allow the current request to
still succeed. The next invocation of the blocking interfaces will then
block. Though, all non-blocking interfaces will still be working and
deliver data. This is consistent with the boot-time behavior of the LRNG
blocking only the "blocking" interfaces.

This change also requires an update to the automated test verifying that
non-blocking interfaces indeed do not block.

Signed-off-by: Stephan Mueller <smueller@chronox.de>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>